fix(core): fix init logging and package.json updates
