+++ packages/nx/src/command-line/init/implementation/angular/index.ts
@@ -8,7 +8,6 @@ import type { PackageJson } from '../../../../utils/package-json';
 import {
   addDepsToPackageJson,
   initCloud,
-  printFinalMessage,
   runInstall,
   updateGitIgnore,
 } from '../utils';
@@ -65,13 +64,6 @@ export async function addNxToAngularCliRepo(options: Options) {
     output.log({ title: '🛠️ Setting up Nx Cloud' });
     initCloud(repoRoot, 'nx-init-angular');
   }
-
-  printFinalMessage({
-    learnMoreLink: 'https://nx.dev/recipes/angular/migration/angular',
-    bodyLines: [
-      '- Execute "npx nx build" twice to see the computation caching in action.',
-    ],
-  });
 }
 
 async function collectCacheableOperations(options: Options): Promise<string[]> {
